+++ packages/admin/src/app/feature/features-menu/features-menu.component.ts
@@ -1,10 +1,5 @@
 import { Component, computed, inject } from '@angular/core';
-import {
-  NavigationEnd,
-  NavigationStart,
-  Router,
-  RouterModule,
-} from '@angular/router';
+import { NavigationEnd, Router, RouterModule } from '@angular/router';
 import { MatIconModule } from '@angular/material/icon';
 import { NgClass } from '@angular/common';
 import { FeatureStore } from '@shared/store/feature.store';
@@ -27,6 +22,7 @@ import {
   faCubes,
   faRectangleList,
   faTags,
+  faUserGroup,
   faGear,
   faUsers,
 } from '@fortawesome/free-solid-svg-icons';
@@ -78,6 +74,7 @@ export class FeaturesMenuComponent {
     faCubes: faCubes,
     faRectangleList: faRectangleList,
     faTags: faTags,
+    faUserGroup: faUserGroup,
     faGear: faGear,
   };