+++ packages/admin/src/app/shared/components/people-table/people-table.component.ts
@@ -1,13 +1,9 @@
 import {
-  AfterViewInit,
   ChangeDetectorRef,
   Component,
-  ViewChild,
-  effect,
   inject,
   input,
   output,
-  viewChild,
 } from '@angular/core';
 import {
   MatCell,