+++ packages/admin/src/app/shared/components/spinner/spinner.component.ts
@@ -1,0 +1,15 @@
+import { CommonModule } from '@angular/common';
+import { Component, inject } from '@angular/core';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { SpinnerService } from '@app/shared/services/spinner.service';
+
+@Component({
+  selector: 'aw-spinner',
+  standalone: true,
+  imports: [CommonModule, MatProgressSpinnerModule],
+  templateUrl: './spinner.component.html',
+  styleUrl: './spinner.component.scss',
+})
+export class SpinnerComponent {
+  readonly spinnerService = inject(SpinnerService);
+}