+++ packages/server/app/models/fact.ts
@@ -9,6 +9,7 @@ import Tag from '#models/tag'
 import Dimension from '#models/dimension'
 import AwBaseModel from '#models/aw_base_model'
 import { ModelQueryBuilderContract } from '@adonisjs/lucid/types/model'
+import Group from '#models/group'
 
 export default class Fact extends AwBaseModel {
   @column({ isPrimary: true })
@@ -60,6 +61,14 @@ export default class Fact extends AwBaseModel {
     pivotRelatedForeignKey: 'tag_id',
   })
   declare tags: ManyToMany<typeof Tag>
+
+  @manyToMany(() => Group, {
+    pivotTimestamps: true,
+    pivotTable: 'group_object',
+    pivotForeignKey: 'object_id',
+    pivotRelatedForeignKey: 'group_id',
+  })
+  declare groups: ManyToMany<typeof Group>
 
   @afterDelete()
   static async detachTags(fact: Fact) {