+++ packages/server/app/models/person.ts
@@ -19,6 +19,7 @@ import Event from '#models/event'
 import AwBaseModel from '#models/aw_base_model'
 import Dimension from '#models/dimension'
 import { ModelQueryBuilderContract } from '@adonisjs/lucid/types/model'
+import Group from '#models/group'
 
 export default class Person extends AwBaseModel {
   @column({ isPrimary: true })
@@ -67,6 +68,14 @@ export default class Person extends AwBaseModel {
     pivotRelatedForeignKey: 'tag_id',
   })
   declare tags: ManyToMany<typeof Tag>
+
+  @manyToMany(() => Group, {
+    pivotTimestamps: true,
+    pivotTable: 'group_object',
+    pivotForeignKey: 'object_id',
+    pivotRelatedForeignKey: 'group_id',
+  })
+  declare groups: ManyToMany<typeof Group>
 
   @afterDelete()
   static async detachTags(person: Person) {