+++ packages/server/database/factories/group_factory.ts
@@ -1,0 +1,11 @@
+import factory from '@adonisjs/lucid/factories'
+import Group from '#models/group'
+
+export const GroupFactory = factory
+  .define(Group, async ({ faker }) => {
+    const name = faker.string.alphanumeric(20)
+    return {
+      name: name,
+    }
+  })
+  .build()